Crate korome

A small game engine written in Rust.

Modules

glium
Re-export of glium crate

Macros

include_texture!
Loads a texture, by loading the bytes at compile-time

is_down!
Macro for easily doing things if particular keys are down

Example

Structs

Drawer
Provides functionality for drawing.

FrameInfo
Wraps together useful data about what has happened (e.g. events)

GameManager
Manages events and frames

Graphics
Contains the display and handles most of the graphics

Texture
A 2D texture that is ready to be drawn

Vector2
Representation of a mathematical vector e.g. a position or velocity

Enums

MouseButton

StateAction
Describes what to do after a frame update

TextureError
Wraps together all errors that can occur creating Textures

VirtualKeyCode

Traits

Draw
Descibes objects that can be drawn to the screen

FloatVector
Special methods for floating point vectors

State
Describes a way to handle an update

Update
Describes objects that can change because of events

Type Definitions

DrawResult
Result type for glium::DrawError

TextureResult
Result type for korome::TextureError

Implementations