Struct korome::GameManager

pub struct GameManager<'a> {
    // some fields omitted
}

Manages events and frames

Methods

impl<'a> GameManager<'a>

fn new(graphics: Graphics<'a>) -> Self

Creates a new GameManager from a Graphics object

fn run(&mut self, initial_state: Box<State>)

Runs the game with the specified initial state

fn run_until_closed<F: FnMut(FrameInfo, Drawer)>(&mut self, f: F)

Runs the game until the user tries to close the window

fn next_frame(&mut self) -> (bool, FrameInfo, Drawer)

Returns a tuple of a bool (which is true if the user tried to close the window), a FrameInfo and a Drawer