Struct korome::Graphics

pub struct Graphics<'a> {
    // some fields omitted
}

Contains the display and handles most of the graphics

Methods

impl<'a> Graphics<'a>

fn new(title: &str, width: u32, height: u32) -> Self

Creates a new Graphics from a Display made using the arguments

fn from_display(display: Display) -> Self

Creates a new Graphics instance using the given display

fn get_h_size(&self) -> (f32, f32)

Returns half of the size of the window